pyproject.toml
[project]
name = "qm-story-engine-2"
version = "0.1.0"
description = "A structured storytelling platform with SQLModel + Supabase backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "sqlmodel>=0.0.21",
    "supabase>=2.8.0",
    "python-dotenv>=1.0.0",
    "pydantic[email]>=2.0.0",
    "pgvector>=0.3.0",
    "openai>=1.50.0",
    "asyncpg>=0.29.0",
    "httpx>=0.26.0",
    "pydantic-settings>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.12.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
]
